refactor(overflow): Add VARCHAR validation and fix pagination integer overflow   - Replace VARCHAR truncation with validation (11 fields across 5 entities).   - Fix critical pagination overflow: int→long offset (11 loops, 8 methods).   - Add safe cast validation for JPA setFirstResult() compatibility.   - Add DATABASE_FIELD_LIMITS.md + update all docs with long offset examples.   - Add 17 comprehensive overflow protection tests.

+++ src/test/java/com/rbatllet/blockchain/performance/PerformanceOptimizationTest.java
@@ -189,7 +189,7 @@ public class PerformanceOptimizationTest {
         long totalBlocks = blockchain.getBlockCount();
         int processedBlocks = 0;
         
-        for (int offset = 0; offset < totalBlocks; offset += BATCH_SIZE) {
+        for (long offset = 0; offset < totalBlocks; offset += BATCH_SIZE) {
             List<Block> batch = blockchain.getBlocksPaginated(offset, BATCH_SIZE);
             processedBlocks += batch.size();
